+++ projects/core/src/main/kotlin/site/siredvin/peripheralworks/computercraft/peripherals/InformativeRegistryPeripheral.kt
@@ -107,6 +107,4 @@ class InformativeRegistryPeripheral(
         result = 31 * result + peripheralOwner.hashCode()
         return result
     }
-
-
 }