+++ projects/core/src/main/kotlin/site/siredvin/peripheralworks/computercraft/peripherals/RecipeRegistryPeripheral.kt
@@ -95,6 +95,4 @@ class RecipeRegistryPeripheral(
         result = 31 * result + peripheralOwner.hashCode()
         return result
     }
-
-
 }