+++ projects/core/src/main/kotlin/site/siredvin/peripheralworks/computercraft/peripherals/RemoteObserverPeripheral.kt
@@ -70,6 +70,4 @@ class RemoteObserverPeripheral(
         result = 31 * result + blockEntity.hashCode()
         return result
     }
-
-
 }