+++ projects/core/src/main/kotlin/site/siredvin/peripheralworks/computercraft/peripherals/StatueWorkbenchPeripheral.kt
@@ -126,6 +126,4 @@ class StatueWorkbenchPeripheral(
         result = 31 * result + peripheralOwner.hashCode()
         return result
     }
-
-
 }