+++ projects/core/src/main/kotlin/site/siredvin/peripheralworks/computercraft/peripherals/turtles/TurtlePeripheraliumHubPeripheral.kt
@@ -116,6 +116,4 @@ class TurtlePeripheraliumHubPeripheral(maxUpdateCount: Int, access: ITurtleAcces
         result = 31 * result + activeTurtleUpgrades.hashCode()
         return result
     }
-
-
 }