+++ projects/fabric/src/main/kotlin/site/siredvin/peripheralworks/integrations/automobility/AutomobilePlugin.kt
@@ -4,7 +4,7 @@ import dan200.computercraft.api.lua.LuaFunction
 import io.github.foundationgames.automobility.entity.AutomobileEntity
 import site.siredvin.peripheralium.api.peripheral.IPeripheralPlugin
 
-class AutomobilePlugin(private val entity: AutomobileEntity): IPeripheralPlugin {
+class AutomobilePlugin(private val entity: AutomobileEntity) : IPeripheralPlugin {
 
     @LuaFunction(mainThread = true)
     fun rotate(delta: Double) {